requirements.txt
Flask
Flask-Cors
Flask-Bcrypt
Flask-JWT-Extended
Flask-SQLAlchemy
python-dotenv
requests
google-auth
google-auth-oauthlib
